refactor: split ci-routing into routing, content-hash, and CLI modules
Separate path classification / job planning from the content-hash skip
protocol and the argv CLI so each concern can change independently.
Keep scripts/ci-routing.ts as the public re-export entry and
import.meta.main owner so workflow and composite-action commands stay
unchanged.

Route scripts/ci-routing/** through the ci_routing lane and universal
content-hash inputs, and add characterization tests for subtree
routing, export surface, and CLI subcommands.
